moving TextQuestionType folder to Features-->components
